+++ src/Braintree/RiskData.cs
@@ -7,6 +7,7 @@ namespace Braintree
         public virtual string id { get; protected set; }
         public virtual string decision { get; protected set; }
         public virtual bool? deviceDataCaptured { get; protected set; }
+        public virtual string fraudServiceProvider { get; protected set; }
 
         public RiskData(NodeWrapper node)
         {
@@ -15,6 +16,7 @@ namespace Braintree
 
             id = node.GetString("id");
             decision = node.GetString("decision");
+            fraudServiceProvider = node.GetString("fraud-service-provider");
             deviceDataCaptured = node.GetBoolean("device-data-captured");
         }